init JSON Processing - CarDealer - p. 17. Export Cars With Their List Of Parts

+++ CSharp-Entity_Framework_Core/JSON-Processing-Exercises/CarDealer-6.0/CarDealer/Models/Car.cs
@@ -8,7 +8,7 @@ public class Car
 
     public string Model { get; set; }
 
-    public int TraveledDistance { get; set; }
+    public long TraveledDistance { get; set; }
 
     public ICollection<PartCar> PartCars { get; set; }
 }
